+++ src/App.css
@@ -28,7 +28,6 @@ h1 {
 header {
   background-color: var(--primary);
 }
-
 .small h1 {
   font-size: 1.25em;
   padding-bottom: 0;
@@ -39,103 +38,14 @@ header {
   font-size: 2em;
 }
 
-.links {
-  display: flex;
-  padding-left: 20px;
-}
-
-.links a {
-  font-size: 0.8em;
-  padding: 6px 10px 20px 0;
-  text-decoration: none;
-  color: white;
-}
-
 section {
   margin: 20px auto;
   padding: 0 20px;
   max-width: 700px;
 }
 
-section#contact {
-  max-width: 500px;
-}
-
 .MuiTabs-root {
   background-color: var(--primary);
-}
-
-#portfolio {
-  display: flex;
-  direction: row;
-  justify-content: center;
-  align-items: start;
-  flex-wrap: wrap;
-}
-
-#portfolio .techs {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: start;
-  align-items: start;
-  margin-bottom: 6px;
-}
-
-#portfolio a.git {
-  text-decoration: none;
-  color: var(--link);
-  font-weight: 600;
-}
-
-#portfolio .MuiPaper-root {
-  margin: 0 auto 20px;
-  background-color: var(--primary-light-bg);
-  color: white;
-  width: 320px;
-}
-
-#portfolio .MuiCardHeader-root {
-  background-color: var(--primary);
-  color: white;
-  padding: 10px;
-}
-
-#portfolio .MuiCardContent-root {
-  padding: 6px 10px 10px;
-}
-
-#portfolio .MuiCardContent-root p {
-  color: white;
-}
-#portfolio .MuiCardContent-root span {
-  color: var(--primary);
-  font-weight: 700;
-  font-size: 0.9em;
-  margin-right: 8px;
-}
-
-#portfolio .MuiButtonBase-root {
-  padding-top: 0;
-}
-
-#portfolio .MuiButtonBase-root svg {
-  color: white;
-}
-
-#contact .MuiFormControl-root {
-  padding: 0;
-  width: 100%;
-  margin: 5px 0;
-}
-
-#materials {
-  text-align: center;
-}
-
-#materials a {
-  width: 250px;
-  display: block;
-  margin: 10px auto;
 }
 
 .MuiBox-root a.MuiButtonBase-root {
@@ -150,14 +60,6 @@ section#contact {
 
 a.MuiButtonBase-root.Mui-selected {
   color: var(--secondary);
-}
-
-img.headshot {
-  width: 100px;
-  float: left;
-  border: 3px solid black;
-  border-radius: 5px;
-  margin: 0 10px 10px 0;
 }
 
 footer {
