+++ src/components/index.tsx
@@ -1,21 +1,6 @@
 import { Header } from "./Header";
-import { About } from "./About";
-import { Contact } from "./Contact";
-import { Portfolio } from "./Portfolio";
-import { Materials } from "./Materials";
 import { Footer } from "./Footer";
 import { ProjectCard } from "./ProjectCard";
-import { Pages } from "./Pages";
 import { TechList } from "./TechList";
 
-export {
-  Header,
-  About,
-  Contact,
-  Portfolio,
-  Materials,
-  Footer,
-  ProjectCard,
-  Pages,
-  TechList,
-};
+export { Header, Footer, ProjectCard, TechList };
