+++ src/pages/About/About.css
@@ -1,0 +1,7 @@
+img.headshot {
+  width: 100px;
+  float: left;
+  border: 3px solid black;
+  border-radius: 5px;
+  margin: 0 10px 10px 0;
+}
