+++ src/pages/About/index.tsx
@@ -1,0 +1,3 @@
+import { About } from "./About";
+
+export { About };
