+++ src/pages/Contact/Contact.css
@@ -1,0 +1,9 @@
+section#contact {
+  max-width: 500px;
+}
+
+#contact .MuiFormControl-root {
+  padding: 0;
+  width: 100%;
+  margin: 5px 0;
+}
