+++ src/pages/Contact/index.tsx
@@ -1,0 +1,3 @@
+import { Contact } from "./Contact";
+
+export { Contact };
