+++ src/pages/Materials/Materials.tsx
@@ -1,3 +1,4 @@
+import "./Materials.css";
 import { useEffect } from "react";
 import Button from "@mui/material/Button";
