+++ src/pages/Materials/index.tsx
@@ -1,0 +1,3 @@
+import { Materials } from "./Materials";
+
+export { Materials };
