+++ src/pages/Portfolio/index.tsx
@@ -1,0 +1,3 @@
+import { Portfolio } from "./Portfolio";
+
+export { Portfolio };
