+++ src/pages/index.tsx
@@ -1,0 +1,6 @@
+import { About } from "./About";
+import { Materials } from "./Materials";
+import { Portfolio } from "./Portfolio";
+import { Contact } from "./Contact";
+
+export { About, Materials, Portfolio, Contact };
